Remove mutexes from shared memory. #470
There is no reasons to store mutexes in shared memory,
because only one application can work with database.
If application was interupted when mutex was in locked state,
next startup of application can't reset locked state.
That is why, after this commit the mutex are storing in
the heap memory.

As I understood, array of mutexes were added for case of
interupting of application when mutex was in locked state.
Unlocking of locked mutexes may be done on open database,
but as described above there are no reasons to store mutexes
in shared memory.

If you're trying to make a big number of requests to
database, then you will get a segfault, when one of a threads
after lock-timeout switches to next mutex, because it will
modify a data, which is already modified by an other thread.
That is why there is only one mutex now.

Lock-timeout is keeped, because it helps to flush locking state,
when many threads trying to lock it by restarting locking order.